fix(camera): immediately discard any unsent messages
Before, even when sending a KeyboardInterrupt, the program would stop on the ZMQ context termination
because it wanted to send the remaining lingering messages. However, this meant that you actually
needed to send two KeyboardInterrupts to stop the program. By setting the socket option for
lingering to False, we now get our intended function of applying a single KeyboardInterrupt.